+++ wow_vanilla_messages/src/world/v1/v12/cmsg_accept_trade.rs
@@ -18,7 +18,7 @@ pub struct CMSG_ACCEPT_TRADE {
 }
 
 impl ClientMessageWrite for CMSG_ACCEPT_TRADE {
-    const OPCODE: u32 = 0x11a;
+    const OPCODE: u16 = 0x11a;
 
     fn size_without_size_or_opcode_fields(&self) -> u16 {
         Self::size() as u16
