+++ wow_vanilla_messages/src/world/v1/v12/cmsg_add_ignore.rs
@@ -17,7 +17,7 @@ pub struct CMSG_ADD_IGNORE {
 }
 
 impl ClientMessageWrite for CMSG_ADD_IGNORE {
-    const OPCODE: u32 = 0x6c;
+    const OPCODE: u16 = 0x6c;
 
     fn size_without_size_or_opcode_fields(&self) -> u16 {
         self.size() as u16
