+++ wow_vanilla_messages/src/world/v1/v12/cmsg_auction_list_items.rs
@@ -27,7 +27,7 @@ pub struct CMSG_AUCTION_LIST_ITEMS {
 }
 
 impl ClientMessageWrite for CMSG_AUCTION_LIST_ITEMS {
-    const OPCODE: u32 = 0x258;
+    const OPCODE: u16 = 0x258;
 
     fn size_without_size_or_opcode_fields(&self) -> u16 {
         self.size() as u16
