+++ wow_vanilla_messages/src/world/v1/v12/cmsg_autoequip_item_slot.rs
@@ -20,7 +20,7 @@ pub struct CMSG_AUTOEQUIP_ITEM_SLOT {
 }
 
 impl ClientMessageWrite for CMSG_AUTOEQUIP_ITEM_SLOT {
-    const OPCODE: u32 = 0x10f;
+    const OPCODE: u16 = 0x10f;
 
     fn size_without_size_or_opcode_fields(&self) -> u16 {
         Self::size() as u16
