+++ wow_vanilla_messages/src/world/v1/v12/cmsg_autostore_bag_item.rs
@@ -20,7 +20,7 @@ pub struct CMSG_AUTOSTORE_BAG_ITEM {
 }
 
 impl ClientMessageWrite for CMSG_AUTOSTORE_BAG_ITEM {
-    const OPCODE: u32 = 0x10b;
+    const OPCODE: u16 = 0x10b;
 
     fn size_without_size_or_opcode_fields(&self) -> u16 {
         Self::size() as u16
