+++ wow_vanilla_messages/src/world/v1/v12/cmsg_banker_activate.rs
@@ -19,7 +19,7 @@ pub struct CMSG_BANKER_ACTIVATE {
 }
 
 impl ClientMessageWrite for CMSG_BANKER_ACTIVATE {
-    const OPCODE: u32 = 0x1b7;
+    const OPCODE: u16 = 0x1b7;
 
     fn size_without_size_or_opcode_fields(&self) -> u16 {
         Self::size() as u16
