+++ wow_vanilla_messages/src/world/v1/v12/cmsg_battlefield_status.rs
@@ -17,7 +17,7 @@ pub struct CMSG_BATTLEFIELD_STATUS {
 }
 
 impl ClientMessageWrite for CMSG_BATTLEFIELD_STATUS {
-    const OPCODE: u32 = 0x2d3;
+    const OPCODE: u16 = 0x2d3;
 
     fn size_without_size_or_opcode_fields(&self) -> u16 {
         Self::size() as u16
