+++ wow_vanilla_messages/src/world/v1/v12/cmsg_battlemaster_hello.rs
@@ -19,7 +19,7 @@ pub struct CMSG_BATTLEMASTER_HELLO {
 }
 
 impl ClientMessageWrite for CMSG_BATTLEMASTER_HELLO {
-    const OPCODE: u32 = 0x2d7;
+    const OPCODE: u16 = 0x2d7;
 
     fn size_without_size_or_opcode_fields(&self) -> u16 {
         Self::size() as u16
