+++ wow_vanilla_messages/src/world/v1/v12/cmsg_busy_trade.rs
@@ -17,7 +17,7 @@ pub struct CMSG_BUSY_TRADE {
 }
 
 impl ClientMessageWrite for CMSG_BUSY_TRADE {
-    const OPCODE: u32 = 0x118;
+    const OPCODE: u16 = 0x118;
 
     fn size_without_size_or_opcode_fields(&self) -> u16 {
         Self::size() as u16
