+++ wow_vanilla_messages/src/world/v1/v12/cmsg_channel_owner.rs
@@ -17,7 +17,7 @@ pub struct CMSG_CHANNEL_OWNER {
 }
 
 impl ClientMessageWrite for CMSG_CHANNEL_OWNER {
-    const OPCODE: u32 = 0x9e;
+    const OPCODE: u16 = 0x9e;
 
     fn size_without_size_or_opcode_fields(&self) -> u16 {
         self.size() as u16
