+++ wow_vanilla_messages/src/world/v1/v12/cmsg_clear_trade_item.rs
@@ -18,7 +18,7 @@ pub struct CMSG_CLEAR_TRADE_ITEM {
 }
 
 impl ClientMessageWrite for CMSG_CLEAR_TRADE_ITEM {
-    const OPCODE: u32 = 0x11e;
+    const OPCODE: u16 = 0x11e;
 
     fn size_without_size_or_opcode_fields(&self) -> u16 {
         Self::size() as u16
