+++ wow_vanilla_messages/src/world/v1/v12/cmsg_creature_query.rs
@@ -20,7 +20,7 @@ pub struct CMSG_CREATURE_QUERY {
 }
 
 impl ClientMessageWrite for CMSG_CREATURE_QUERY {
-    const OPCODE: u32 = 0x60;
+    const OPCODE: u16 = 0x60;
 
     fn size_without_size_or_opcode_fields(&self) -> u16 {
         Self::size() as u16
