+++ wow_vanilla_messages/src/world/v1/v12/cmsg_del_ignore.rs
@@ -19,7 +19,7 @@ pub struct CMSG_DEL_IGNORE {
 }
 
 impl ClientMessageWrite for CMSG_DEL_IGNORE {
-    const OPCODE: u32 = 0x6d;
+    const OPCODE: u16 = 0x6d;
 
     fn size_without_size_or_opcode_fields(&self) -> u16 {
         Self::size() as u16
